Refactor: 카카오 API 함수 분리 및 이름 수정
함수가 하나의 일을 하도록 분리
