+++ src/test/java/com/ssafy/tott/api/kakao/module/KakaoAPITest.java
@@ -1,6 +1,7 @@
 package com.ssafy.tott.api.kakao.module;
 
-import com.ssafy.tott.api.kakao.data.Documents;
+import com.ssafy.tott.api.kakao.data.vo.Documents;
+import com.ssafy.tott.api.kakao.service.KakaoMapFetchAPI;
 import com.ssafy.tott.api.seoulopendata.data.vo.RentRow;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
@@ -21,12 +22,12 @@ class KakaoAPITest {
     private final RentRow row = new RentRow("2023", "11380", "은평구", "10300", "불광동", "1", "대지", "0105", "0076", 3.0, "20230901", "전세", 57.76, "23000", "0", "105-76", "2018", "연립다세대", "", "신규", "", "0", "");
 
     @Autowired
-    private KakaoMapAPI kakaoMapAPI;
+    private KakaoMapFetchAPI kakaoMapFetchAPI;
 
     @DisplayName("kakaoAPI")
     @Test
     void kakaoAPIConnectTest() {
-        Documents response = kakaoMapAPI.kakaoAddressSearch(row.getSggNm(), row.getBjdongNm(), row.getBobn(), row.getBubn());
+        Documents response = kakaoMapFetchAPI.kakaoAddressSearch(row.getSggNm(), row.getBjdongNm(), row.getBobn(), row.getBubn());
         log.info(String.valueOf(response));
 
         assertThat(response).isNotNull();
